PIPELINE-DASH/dags/requirements.txt
apache-airflow
google-cloud-bigquery
requests
